App/requirements.txt
Flask~=2.1.1
python-dotenv~=0.20.0
mercantile~=1.2.1
Pillow==9.1.0
furl~=2.1.3
requests~=2.27.1
click==8.1.3
torch~=1.11.0
torchvision~=0.12.0
opencv-python~=4.5.5.64
numpy==1.22.3
albumentations~=1.1.0
datetime~=4.4